move instantiateWasm
