+++ src/native/corehost/browserhost/host/per-module.ts
@@ -1,0 +1,4 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+export * from "../../../libs/Common/JavaScript/per-module";
